Update SimpleAnonymizer.py
Fixed an issue with initials that were also single letter words. For example, "This is a word" would be anonymized into "This is a NAME word" for a person with the name "Robert A Wilson." Now, if an initial is detected it the prev and next words must also be a name.